Fix bug with X11 in MacOS
The issue was in the subprocess.call for 'xhost +',
where the permission granted seem to vanish somehow.
As I have done for the 'docker run' call, changed
subprocess.call for 'os.system' and voila'!
There was also a minor issue with bytearrays, the
host machine address was being printed with the "b''"
in front of the IP (only in Python3).